updated things after my Santa Barbara visit
Most important change: we've decided that we want to be finding
galaxies with a stellar mass between 10^8 and 10^9, not just
any galaxy below 10^9 (which is what we were previously doing).
This change was to help seperate the target population from the
general population in magnitude space. The low mass dwarfs were
making it difficult to tell the difference between dim, nearby
galaxies, and more luminous faraway galaxies, but they wouldn't
add much to the lensing signal. So by cutting them out, it
becomes easier to identify good dwarfs without (hopefully)
losing too much lensing signal.

A bunch of smaller updates have also been thrown in.